Done IG

+++ app/src/main/java/com/example/eventplanner/View/Fragments/CategoryListFragment.kt
@@ -5,29 +5,23 @@ import android.content.ContentValues
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
+import android.view.*
 import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.eventplanner.R
 import com.example.eventplanner.View.Activities.AddCategoryActivity
-import com.example.eventplanner.View.Activities.AddEventActivity
 import com.example.eventplanner.View.Adapters.CategoryListAdapter
-import com.example.eventplanner.View.Adapters.DoneAdapter
-import com.example.eventplanner.View.Fragments.BottomSheet.BottomSheetFragment
 import com.example.eventplanner.View.Fragments.BottomSheet.CategoryBottomSheetFragment
 import com.example.eventplanner.viewModel.CategoryViewModel
-import com.example.eventplanner.viewModel.ListViewModel
+import com.example.eventplanner.viewModel.parcels.CategoriesWithEvent
 import com.example.eventplanner.viewModel.parcels.Category
-import com.example.eventplanner.viewModel.parcels.Event
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 
 class CategoryListFragment : Fragment(R.layout.fragment_category_list) {
-    private val TAG = "CalendarListFragment"
+    private val TAG = "CategoryListFragment"
     private lateinit var recyclerView: RecyclerView
     private lateinit var adapter: CategoryListAdapter
     private lateinit var viewModel : CategoryViewModel
@@ -63,14 +57,36 @@ class CategoryListFragment : Fragment(R.layout.fragment_category_list) {
         bottomSheet.show(childFragmentManager, "Long")
     }
 
+    override fun onContextItemSelected(item: MenuItem): Boolean {
+        return when (item.itemId){
+            101 -> {
+               editCategory(adapter.getCategory(item.groupId))
+                true
+            }
+            102 -> {
+                deleteCategory(adapter.getCategory(item.groupId))
+                true
+            }
+            else -> super.onContextItemSelected(item)
+        }
+    }
 
-    fun editCategory(category: Category){
+
+    private fun editCategory(category: Category){
         categoryActivityIntent = Intent(context, AddCategoryActivity::class.java)
         categoryActivityIntent!!.putExtra("category", category)
         getCategory.launch(categoryActivityIntent)
     }
 
-    fun deleteCategory(category: Category){
-        viewModel.deleteCategory(category, requireView())
+    private fun deleteCategory(category: Category){
+        MaterialAlertDialogBuilder(requireContext())
+            .setTitle("Warning")
+            .setMessage("You will lose the data permanently\nAre you sure?")
+            .setNegativeButton("No"){ dialog, which->
+
+            }
+            .setPositiveButton("Yes"){dialog, which->
+                viewModel.deleteCategory(category, requireView())
+            }.show()
     }
 }
